Always inhibit operator code when canceling an exchange.
When `evil-exchange` is triggered through the `cx` keybinding, and
then canceled with `c`, this function runs and eventually evil treats
the result as a trivial motion to pass through to the rest of the
operator code.

Since we're canceling the exchange, we never want the rest of the
`evil-exchange` code to run, so we have to explicitly cancel it here.